refactor: move Movie, Screening to view package
