+++ app/src/main/java/woowacourse/movie/view/MovieAdapter.kt
@@ -8,7 +8,7 @@ import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
 import woowacourse.movie.R
-import woowacourse.movie.domain.Screening
+import woowacourse.movie.view.model.Screening
 
 class MovieAdapter(
     screenings: List<Screening>,
